Fix javadoc code error

+++ app/src/main/java/azul/Row.java
@@ -2,13 +2,13 @@ package azul;
 
 public class Row {
     private Color color;
-    private byte number;
+    private int number;
 
     public Color getColor() {
         return color;
     }
 
-    public byte getNumber() {
+    public int getNumber() {
         return number;
     }
 }
